+++ src/main/kotlin/minesweeper/domain/Coordinate.kt
@@ -1,0 +1,8 @@
+package minesweeper.domain
+
+class Coordinate(private val index: Int, private val matrix: Matrix) {
+    val sideIndexes: List<Int>
+        get() {
+            return matrix.around(index).sorted()
+        }
+}
